version 1.0.6
Bugfix:
When transitioning from chess.com/play to chess.com/game the pageClickListener would not get injected.
This is because the change happens dynamically, so no load of the tab.
To circumvent this, pageClickListener will be injected for all chess.com pages.

Bugfix:
Eval lines wouldn't show when the tab was not active.
Now we wait for the tab to be active before clicking on ceval.